Update update-script.rb
Organisations using the old org.visualstudio.com URL rather than the new dev.azure.com/org structure don't have the same repo path. This allows them to override this fully if the organisation is not yet in a state to migrate domains.

If the repo_name is provided in the pipeline inputs as standard (eg. `dependabot-azure-devops`) then it will behave as standard and build the URL using the other variables.

If it is provided as a full URL (eg. `https://org.visualstudio.com/Project/_git/dependabot-azure-devops`) then this will be used as-provided.