Support/Option to skip creating and updating all PRs (#419)
This PR adds support to skip creating and updating all PRs for updates. This is particularly useful for debugging without the creation of PRs and for PoC scenarios. This option is disabled by default.

To enable in Docker pass `-e DEPENDABOT_SKIP_PULL_REQUESTS=true`

To enable in the pipeline:

```yaml
steps:
- task: dependabot@1
  inputs:
    skipPullRequests: true
```