Pull security vulnerabilities from GitHub's GraphQL endpoint (#434)
This is a continuation of #430 but now introduces support for querying vulnerabilities from the GraphQL endpoint just for the given ecosystem and package hence smaller payload bodies and no longer need to maintain local data. No additional setup is required except for a GitHub access token with `public_repo` scope.