Image per ecosystem (#720)
This PR moves from the single image which is now dated, on to having an image per ecosystem. This alignes with the GitHub hosted version.

Some more information available [here](https://github.com/dependabot/dependabot-script/issues/896#issuecomment-1425382734)

As a consequence, the image repository can no longer be specified as an input to the task (using `dockerImageRepository`) or the server (using `updaterImageRepository`).