Update the server version/component to support directories and groups
The extension already supports the directories and groups using `update_script_vnext` but the server lies behind. This standardizes support.

The image tag override for NuGet is removed from the bicep resources because the `1.24` image no longer exists and because we no longer need to override it. If you had deployed the app configuration instance, either manually remove the configuration entry from the Azure portal or redeploy using complete mode (defaults to incremental which does not delete).