Add prettier formatting
This will mostly format files in the extension but can format some other non-js/ts files that it supports like json, yaml, etc.

At the root `package.json` and `package-lock.json` files are added but a workspace is not used because we do not need it and it is preferred for the extension to be independent to ensure it is self-contained during deployment. This may change if we can use `tsup` to package it.