Make masking of secrets optional via `System.Secrets` ENV (defaults to false)
Apparently, lots of people believe that since they work in a corporate repo, even the URLs are very secret. While it makes sense in some cases, it more often than not is getting in the way of being able to debug correctly.
A new environment variable, `System.Secrets` that is not official but used in this PR, makes it an opt-in functionality.