No longer storing dependency list in project properties
The list of dependencies can grow quite large (>10KB in non trivial repos) and may surpass the limit allowed or slow down the AzDo UI which loads all properties.
This PR removes that code and outputs an warning instead as we research if there is a valid alternative.

Consequently, all code that access/modifies project properties is also removed as it is unused.