Support deploying to local docker
This is a simple way to test but it is also useful incase one wants to run on a Raspberry PI or outside Azure though I wonder who the latter guy is.

In this mode, registration of webhooks does not make sense so it is now optional.

Other deployment targets can be added later by the community if needed.