Removing task v1 and updater
The old updater images which we build upon are starting to be removed or to disappear. There are users still using this and there is so much that has changed since which may result in too much maintenance/support work.
Running with v1 task will result in an error.

We will no longer produce `ghcr.io/tinglesoftware/dependabot-updater-{ecosystem]` images though those present will continue to be there for a little longer incase one wants to execute the image directly. The server still uses these images but will soon move to the same workflow as task v2.