Remove support for reviewers (assignees remain)
This aligns with support for this in deprecation announcement about 2-3 weeks
https://github.blog/changelog/2025-04-29-dependabot-reviewers-configuration-option-being-replaced-by-code-owners/

We have always joined assignees and reviewers because Azure DevOps does not have the concept of assignees.

GitHub is deprecating reviewers in favour of CODEOWNERS. This is also something that Azure DevOps does not support but it can be done using Branch Policies.