Restore task v1
- Revert "Removing task v1 and updater ([#1660](https://github.com/tinglesoftware/dependabot-azure-devops/issues/1660))"
  This reverts commit 0c937c36fbb98fb9bea1f4cd3acd080f28f95494.
- Revert "Remove support for `securityAdvisoriesFile` ([#1661](https://github.com/tinglesoftware/dependabot-azure-devops/issues/1661))"
  This reverts commit 5ac0d015a2685ccfe3bc967d71a086031e348a24.

Why? There seems to be a huge number of people who are yet to migrate. A few due to the lack of custom security advisories. While that is being worked on and the server is prepare to v2, we allow the use of v1.
Initially, there was no other way to tell if there are people affected by the drop.