Bundle extension with tsup (#1698)
As we add tasks, we tend to duplicate logic and also copying of node_modules but this makes the vsix bulky and we reduce code sharing.
By using a bundler (in this case, tsup), we can share logic then bundle. This should reduce the extension size drastically (less than 2MB). Smaller size means faster uploads and less failures when uploading to the store.

PNPM is now the default package manager in use which gives us lots of freedom.