Support for private security vulnerabilities
Added support for passing vulnerabilities via a file, usually for private dependencies. To conform to the what we expect from GitHub, some logic is reused and the sample JSON file is updated.
Zod is used to validate that the contents of the file match the expectation.

This was supported in v1 of the task extension but since that was removed, this functionality was missing in v2.