Added a new core package which stores some logic to be shared with zod validation (#1760)
This logic is (to) shared by extensions, dashboard, and servers and helps standardize how things will run going forward.

Key addition is the use of zod to parse the config. This reduces lots of validation  boilerplate and ensures we are consistent with typing.

It also allows the creation of a simple CLI tool that can be used as an alternative to the extension.