Treat assignees as optional reviewers
With the removal of `reviewers` in #1688, the recommendation is to use CODEOWNERS which GitHub uses for all PRs.

In place of CODEOWNERS, Azure DevOps offers branch policies to add required reviewers and this is the recommendation. However, Azure DevOps still does not have the concept of `assignees` and we have had to merge those into `reviewers` by treating them as required reviewers. This no longer aligns semantically with the logic that GitHub uses. I could not fina any officially clear separation of reviewers and assignees but this [reply](https://stackoverflow.com/a/41174022) explains it quite well.

With this PR, assignees are no longer required reviewers. To add required reviewers, one should use branch policies which should already be in place in most cases.