Fix versioning of extension and task in dev
This used to work but when after moving to changesets, publishing dev extensions did not work.
With this the build number will be added to the extension version at the end (4th digit). The patch field for the task will be replaced with the build number during dev (anything but tagged).