Collect affected PRs for a given run and set output variable
To allow other tasks to be chained to be placed after dependabot runs, it makes sense to collect the identifiers of pull requests created/modified/abandoned then output these.
This change introduces a comma-separated value is set to a new `affectedPrs` output variable.

It should not be easier to do any other logic that one needs to do such as making reviewers required, sending notifications to Teams, or light up some office lights.