Added CLI with command to validate a dependabot configuration file (#1796)
Introduces a new CLI with a command that can validate a Dependabot configuration file. This can help ensure the config is well-formed and adheres to expected structure before it's committed or used.

The CLI is being added a simple solution that one can use to automate in different workflows. It is aimed at anyone running outside Azure Pipelines (e.g. on a Raspberry PI) or scenarios where people used the earlier docker images to run updates.

This CLI might also end up replacing the use of [`dependabot-cli`](https://github.com/dependabot/cli) with direct usage of docker as is done in [`dependabot-action`](https://github.com/github/dependabot-action).

> The `generate` command present at this time is not yet fully functional. That will happen in a followup PR/commit